refactor(gax): remove duplicate StaticTraceContext and export from TracerHelper
