+++ core/packages/gax/src/fallback.ts
@@ -31,7 +31,8 @@ import * as fallbackRest from './fallbackRest';
 import {isNodeJS} from './featureDetection';
 import {generateServiceStub} from './fallbackServiceStub';
 import {StreamType} from './streamingCalls/streaming';
-import {toLowerCamelCase, StaticTraceContext} from './util';
+import {toLowerCamelCase} from './util';
+import {StaticTraceContext} from './observability/TracerHelper';
 import {google} from '../protos/http';
 import * as IamProtos from '../protos/iam_service';
 import * as LocationProtos from '../protos/locations';
